requirements.txt
psycopg2-binary==2.9.1
requests==2.25.1
#jsonlib-python3==1.6.1